[MAR-2640] Close second-round review findings
